Fix version number and mistakes in README.md
I merged in all outstanding pull requests, but since most of them
changed README.md in some way I had to do a lot of conflict resolution
and ended up messing up some of the indentation.  This fixes that, as
well as upping the version number in the installation instructions to
match what's in the latest git-media.gemspec.